firmware/utils/requirements.txt
# для приложения
wxPython
pyserial
# для симуляции
numpy
pyQt5
